requirements.txt
PyQt5==5.15.9
python-dotenv==1.0.0
requests==2.28.2
